Refactor: modulize functions
