+++ js/signUtils.js
@@ -1,0 +1,70 @@
+const userID = document.querySelector("#email");
+const userPW = document.querySelector("#password");
+const emailError = document.querySelector(".emailError");
+const pwError = document.querySelector(".pwError");
+const pwContainer = document.querySelector(".pw-container");
+
+export function errorStyle(textStyle, boxStyle) {
+  textStyle.style.color = "red";
+  boxStyle.style.border = "1px solid red";
+}
+
+export function clearStyle(textStyle, boxStyle) {
+  textStyle.textContent = "";
+  boxStyle.style.border = "";
+}
+
+export function validEmail(obj) {
+  if (obj.value === "") {
+    emailError.textContent = "이메일을 입력해주세요.";
+    errorStyle(emailError, userID);
+  } else if (validEmailCheck(obj) == false) {
+    emailError.textContent = "올바른 이메일 주소가 아닙니다.";
+    errorStyle(emailError, userID);
+  } else {
+    clearStyle(emailError, userID);
+  }
+}
+
+export function validPassword(obj) {
+  if (obj.value === "") {
+    pwError.textContent = "비밀번호를 입력해주세요.";
+    errorStyle(pwError, userPW);
+  } else if (validPasswordCheck(obj) == false) {
+    pwError.textContent = "비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요.";
+    errorStyle(pwError, userPW);
+  } else {
+    clearStyle(pwError, userPW);
+  }
+}
+
+export function goUrl() {
+  let link = "folder/";
+  location.href = link;
+}
+
+export function validEmailCheck(obj) {
+  const pattern =
+    /^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}$/i;
+  return !!obj.value.match(pattern);
+}
+
+export function validPasswordCheck(obj) {
+  const pattern = /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,25}$/;
+  return !!obj.value.match(pattern);
+}
+
+export function showPW(e) {
+  const lineThrough = this.querySelector(".line-through");
+  const pwInput = this.querySelector("#password");
+  if (e.target.classList.contains("eye-PW"))
+    if (pwInput.type === "password") {
+      pwInput.type = "text";
+      lineThrough.style.display = "block";
+    } else {
+      pwInput.type = "password";
+      lineThrough.style.display = "none";
+    }
+}
+
+export { userID, userPW, emailError, pwError, pwContainer };
